requirements.txt
pyyaml
tweepy
